requirements.txt
Django>=4.0.0
Pillow>=9.0.0
torch>=1.12.0
torchvision>=0.13.0
pymongo>=4.0.0
